safari-600.7-branch/LayoutTests/fast/dom/Window/window-screen-properties-expected.txt

PASS: window.screen.pixelDepth is either 16 or 24
PASS: window.screen.colorDepth is either 16 or 24
